refactor: move unbundled regenerator test to build-options
- I accidentally put it into build-default before build-options existed,
  but the test very specifically tests an option, `--target node`, so
  it should be in build-options as that's not a zero-config default
